調整預存程序，因此需要更新 EDMX 模型

+++ MVC5Demo/Models/GetCourseReport_Result.cs
@@ -17,6 +17,6 @@ namespace MVC5Demo.Models
         public string CourseName { get; set; }
         public Nullable<int> TeacherCount { get; set; }
         public Nullable<int> StudentCount { get; set; }
-        public Nullable<int> AvgGrade { get; set; }
+        public Nullable<double> AvgGrade { get; set; }
     }
 }
